requirements.txt
openai>=1.30.0
boto3>=1.34.0
anthropic>=0.28.0
langchain>=0.2.0
langgraph>=0.1.0
sentence-transformers>=3.0.0
numpy>=1.26.0
scipy>=1.12.0
pandas>=2.2.0
matplotlib>=3.8.0
seaborn>=0.13.0
pyyaml>=6.0
pytest>=8.0.0
tqdm>=4.66.0
